docker/requirements.txt
torch==2.0.1
rdkit-pypi==2022.9.5
transformers==4.31.0
pandas==2.0.0
numpy==1.24.2
tqdm==4.66.1
h5py==3.10.0
wandb==0.15.0
pytorch-lightning==2.3.2
torchmetrics==1.4.0.post0
argparse==1.4.0
FuzzyTM>=0.4.0
blosc2~=2.0.0
platformdirs>=2.1